feat(nimbus): rework experiment review checks into its own hook, make refetch top-level
