+++ app/experimenter/nimbus-ui/src/components/FormOverview/mocks.tsx
@@ -9,7 +9,6 @@ import { MockedCache } from "../../lib/mocks";
 export const Subject = ({
   isLoading = false,
   isServerValid = true,
-  isMissingField = () => false,
   submitErrors = {},
   onSubmit = () => {},
   onCancel,
@@ -25,7 +24,6 @@ export const Subject = ({
         {...{
           isLoading,
           isServerValid,
-          isMissingField,
           setSubmitErrors,
           onSubmit,
           onCancel,
