+++ app/experimenter/nimbus-ui/src/components/PageEditBranches/index.tsx
@@ -83,9 +83,9 @@ const PageEditBranches: React.FunctionComponent<RouteComponentProps> = () => {
       testId="PageEditBranches"
       redirect={editCommonRedirects}
     >
-      {({ experiment, review }) => {
+      {({ experiment, refetch }) => {
         currentExperiment.current = experiment;
-        refetchReview.current = review.refetch;
+        refetchReview.current = refetch;
 
         const applicationFeatureConfigs =
           featureConfig?.filter(
