+++ app/experimenter/nimbus-ui/src/hooks/index.ts
@@ -12,3 +12,4 @@ export * from "./useExitWarning";
 export * from "./useExperiment";
 export * from "./useFakeMutation";
 export * from "./useOutcomes";
+export * from "./useReviewCheck";
